Move feed module out of options module
